+++ src/main/java/com/tsp/new_tsp_front/api/model/service/impl/FrontModelJpaRepository.java
@@ -6,7 +6,6 @@ import com.querydsl.jpa.impl.JPAQueryFactory;
 import com.tsp.new_tsp_front.api.model.domain.FrontModelDTO;
 import com.tsp.new_tsp_front.api.model.domain.FrontModelEntity;
 import com.tsp.new_tsp_front.common.utils.StringUtil;
-import com.tsp.new_tsp_front.exception.TspException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Repository;
@@ -18,8 +17,6 @@ import static com.tsp.new_tsp_front.api.common.domain.QCommonImageEntity.commonI
 import static com.tsp.new_tsp_front.api.model.domain.QFrontModelEntity.frontModelEntity;
 import static com.tsp.new_tsp_front.api.model.service.impl.ModelMapper.INSTANCE;
 import static com.tsp.new_tsp_front.common.utils.StringUtil.getInt;
-import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_MODEL;
-import static com.tsp.new_tsp_front.exception.ApiExceptionType.NOT_FOUND_MODEL_LIST;
 
 @Slf4j
 @RequiredArgsConstructor
@@ -59,27 +56,22 @@ public class FrontModelJpaRepository {
      * 5. 작성일       : 2022. 03. 27.
      * </pre>
      */
-    public List<FrontModelDTO> getMainModelList() throws TspException {
+    public List<FrontModelDTO> getMainModelList() {
+        List<FrontModelEntity> modelList = queryFactory
+                .selectFrom(frontModelEntity)
+                .orderBy(frontModelEntity.idx.desc())
+                .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
+                .fetchJoin()
+                .where(frontModelEntity.modelMainYn.eq("Y")
+                .and(frontModelEntity.visible.eq("Y")
+                .and(commonImageEntity.typeName.eq("model")
+                .and(commonImageEntity.imageType.eq("main"))
+                .and(commonImageEntity.visible.eq("Y")))))
+                .fetch();
 
-        try {
-            List<FrontModelEntity> modelList = queryFactory
-                    .selectFrom(frontModelEntity)
-                    .orderBy(frontModelEntity.idx.desc())
-                    .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
-                    .fetchJoin()
-                    .where(frontModelEntity.modelMainYn.eq("Y")
-                    .and(frontModelEntity.visible.eq("Y")
-                    .and(commonImageEntity.typeName.eq("model")
-                    .and(commonImageEntity.imageType.eq("main"))
-                    .and(commonImageEntity.visible.eq("Y")))))
-                    .fetch();
+        modelList.forEach(list -> modelList.get(modelList.indexOf(list)).setRnum(modelList.indexOf(list)));
 
-            modelList.forEach(list -> modelList.get(modelList.indexOf(list)).setRnum(modelList.indexOf(list)));
-
-            return INSTANCE.toDtoList(modelList);
-        } catch (Exception e) {
-            throw new TspException(NOT_FOUND_MODEL_LIST, e);
-        }
+        return INSTANCE.toDtoList(modelList);
     }
 
     /**
@@ -91,16 +83,12 @@ public class FrontModelJpaRepository {
      * 5. 작성일       : 2022. 03. 27.
      * </pre>
      */
-    public int getModelCount(Map<String, Object> modelMap) throws TspException {
-        try {
-            return queryFactory
-                    .selectFrom(frontModelEntity)
-                    .where(searchModel(modelMap)
-                    .and(frontModelEntity.visible.eq("Y")))
-                    .fetch().size();
-        } catch (Exception e) {
-            throw new TspException(NOT_FOUND_MODEL_LIST, e);
-        }
+    public int getModelCount(Map<String, Object> modelMap) {
+        return queryFactory
+                .selectFrom(frontModelEntity)
+                .where(searchModel(modelMap)
+                .and(frontModelEntity.visible.eq("Y")))
+                .fetch().size();
     }
 
     /**
@@ -112,25 +100,21 @@ public class FrontModelJpaRepository {
      * 5. 작성일       : 2022. 01. 02.
      * </pre>
      */
-    public List<FrontModelDTO> getModelList(Map<String, Object> modelMap) throws TspException {
-        try {
-            List<FrontModelEntity> modelList = queryFactory
-                    .selectFrom(frontModelEntity)
-                    .orderBy(frontModelEntity.idx.desc())
-                    .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
-                    .fetchJoin()
-                    .where(searchModel(modelMap).and(frontModelEntity.visible.eq("Y")))
-                    .offset(getInt(modelMap.get("jpaStartPage"), 0))
-                    .limit(getInt(modelMap.get("size"), 0))
-                    .fetch();
+    public List<FrontModelDTO> getModelList(Map<String, Object> modelMap) {
+        List<FrontModelEntity> modelList = queryFactory
+                .selectFrom(frontModelEntity)
+                .orderBy(frontModelEntity.idx.desc())
+                .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
+                .fetchJoin()
+                .where(searchModel(modelMap).and(frontModelEntity.visible.eq("Y")))
+                .offset(getInt(modelMap.get("jpaStartPage"), 0))
+                .limit(getInt(modelMap.get("size"), 0))
+                .fetch();
 
-            modelList.forEach(list -> modelList.get(modelList.indexOf(list))
-                    .setRnum(getInt(modelMap.get("startPage"),1)*(getInt(modelMap.get("size"),1))-(2-modelList.indexOf(list))));
+        modelList.forEach(list -> modelList.get(modelList.indexOf(list))
+                .setRnum(getInt(modelMap.get("startPage"),1)*(getInt(modelMap.get("size"),1))-(2-modelList.indexOf(list))));
 
-            return INSTANCE.toDtoList(modelList);
-        } catch (Exception e) {
-            throw new TspException(NOT_FOUND_MODEL_LIST, e);
-        }
+        return INSTANCE.toDtoList(modelList);
     }
 
     /**
@@ -142,21 +126,17 @@ public class FrontModelJpaRepository {
      * 5. 작성일       : 2022. 01. 09.
      * </pre>
      */
-    public FrontModelDTO getModelInfo(FrontModelEntity existFrontModelEntity) throws TspException {
-        try {
-            //모델 상세 조회
-            FrontModelEntity getModelInfo = queryFactory
-                    .selectFrom(frontModelEntity)
-                    .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
-                    .fetchJoin()
-                    .where(frontModelEntity.idx.eq(existFrontModelEntity.getIdx())
-                    .and(frontModelEntity.visible.eq("Y"))
-                    .and(commonImageEntity.typeName.eq("model")))
-                    .fetchOne();
+    public FrontModelDTO getModelInfo(FrontModelEntity existFrontModelEntity) {
+        //모델 상세 조회
+        FrontModelEntity getModelInfo = queryFactory
+                .selectFrom(frontModelEntity)
+                .leftJoin(frontModelEntity.commonImageEntityList, commonImageEntity)
+                .fetchJoin()
+                .where(frontModelEntity.idx.eq(existFrontModelEntity.getIdx())
+                .and(frontModelEntity.visible.eq("Y"))
+                .and(commonImageEntity.typeName.eq("model")))
+                .fetchOne();
 
-            return INSTANCE.toDto(getModelInfo);
-        } catch (Exception e) {
-            throw new TspException(NOT_FOUND_MODEL, e);
-        }
+        return INSTANCE.toDto(getModelInfo);
     }
 }
